DBZ-4236 Unify Oracle LogMiner event processors
